requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
blinker==1.7.0
cachelib==0.10.2
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
distro==1.8.0
exceptiongroup==1.1.3
flask==3.0.0
Flask-Cors==4.0.0
flask-session==0.5.0
frozenlist==1.4.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.4
numpy==1.24.4
openai==0.28.1
pandas==2.0.3
pydantic==2.5.1
pydantic-core==2.14.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
tqdm==4.66.1
typing-extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
werkzeug==3.0.1
yarl==1.9.2
zipp==3.17.0
